Deduplicate implementations of LineColumn
